오늘자 백엔드 작업 - 이재욱

+++ artauction/src/main/java/com/art/dto/MemberDto.java
@@ -19,6 +19,6 @@ public class MemberDto {
 	private String memberAddress2;
 	@JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss", timezone = "Asia/Seoul")
 	private Date memberJoinDate;
-	private int memberPoint;
+	private long memberPoint;
 	private boolean isBlocked;
 }
